libraries/Fix behaviour with case sensitivity.
Previously, non-downcase input would not update the match-data properly.
This is because match-data was computed just one, on initialization, before any
input is given.

This is wrong, we must obviously recompute the match-data on every `update'.
Most importantly, we must leave full control the match-data to the filter
functions.  Indeed, it does not have to be just strings.

To that add, we introduce ensure-match-data-string which is called from our
default filter functions when they need it.

Instead of calculating if the input is upcased of downcased for each
suggestion (it could be slow since it's a performance bottleneck), we do it only
when `update' is called and store the result in current-input-downcase-p.

Finally:

- The automatic match-data initialization is removed from `initialize-instance'
and the dedicated `match-data' reader, since NIL is actually an acceptable
value.

- `filter' functions follow the same argument logic as `filter-preprocessor'
and `filter-postprocessor', so they now take (SUGGESTION SOURCE INPUT).

- `to-unicode' has been renamed to the more appropriate
`ensure-non-base-string'.  It is called only when needed:
  - Once per `update' for `input'.
  - When setting match-data in `ensure-match-data-string'.